Back-end/requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.10.0
CacheControl==0.14.4
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
Cython==3.1.6
decorator==5.2.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
easydict==1.13
firebase_admin==7.1.0
flatbuffers==25.9.23
fonttools==4.60.1
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.8.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
insightface==0.7.3
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.7
ml_dtypes==0.5.3
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.2
networkx==3.5
numpy==2.2.6
onnx==1.19.1
onnxruntime==1.23.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
prettytable==3.16.0
proglog==0.1.12
proto-plus==1.27.0
protobuf==6.33.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
simsimd==6.5.3
six==1.17.0
sqlparse==0.5.3
stringzilla==4.2.3
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
whitenoise==6.11.0
